Egoblur integration (#1)
* branch created

* ```download_egoblur_model()``` added

* update config and dev requirements

* duplicate requirements

* add obj name and fix ValueError

* convert boxes to coco

* typeerror fix test

* class name fixed + applicable area

* device logger warning

* get_device() function -> global variable

* slight modal wording fixes

* test both detectors

* "both mode" fix test

* fix obj_classes for both mode

* readme update

* acknowledgement hyperlink added

* config app name change

* add check to avoid warning when unneeded

* faces -> objects

* added docker

* fix model_path type error

* Update config.json

* docker update

---------

+++ src/utils.py
@@ -1,25 +1,44 @@
 import os
 from pathlib import Path
 from typing import Callable, List
+from zipfile import ZipFile
 
 import cv2
 import numpy as np
 import requests
 import supervisely as sly
+import torch
+import torchvision
 
 import globals as g
 
 
+def get_obj_class_names():
+    if g.STATE.target == g.Model.BOTH:
+        return [g.FACE_CLASS_NAME, g.LP_CLASS_NAME]
+    elif g.STATE.target == g.Model.YUNET:
+        return [g.FACE_CLASS_NAME]
+    elif g.STATE.target == g.Model.EGOBLUR:
+        return [g.LP_CLASS_NAME]
+
+
 def updated_project_meta(project_meta: sly.ProjectMeta) -> sly.ProjectMeta:
-    """Update project meta to include anonymized faces"""
+    """Update project meta to include anonymized objects"""
     obj_classes = project_meta.obj_classes
-    if not obj_classes.has_key(g.FACE_CLASS_NAME):
-        obj_classes = obj_classes.add(sly.ObjClass(g.FACE_CLASS_NAME, sly.Rectangle))
-        project_meta = project_meta.clone(obj_classes=obj_classes)
+
+    obj_class_names = get_obj_class_names()
+    for obj_class_name in obj_class_names:
+        if not obj_classes.has_key(obj_class_name):
+            obj_classes = obj_classes.add(sly.ObjClass(obj_class_name, sly.Rectangle))
+    project_meta = project_meta.clone(obj_classes=obj_classes)
     tag_metas = project_meta.tag_metas
     if not g.CONFIDENCE_TAG_META_NAME in tag_metas:
         tag_metas = tag_metas.add(
-            sly.TagMeta(g.CONFIDENCE_TAG_META_NAME, sly.TagValueType.ANY_NUMBER)
+            sly.TagMeta(
+                g.CONFIDENCE_TAG_META_NAME,
+                sly.TagValueType.ANY_NUMBER,
+                applicable_to=sly.TagApplicableTo.OBJECTS_ONLY,
+            )
         )
         project_meta = project_meta.clone(tag_metas=tag_metas)
     return project_meta
@@ -47,7 +66,7 @@ def create_dst_project(src_project: sly.ProjectInfo) -> sly.ProjectInfo:
     g.Api.project.update_meta(dst_project.id, dst_project_meta)
     dst_custom_data = {
         **src_project.custom_data,
-        "anonymized_faces": True,
+        "anonymized_objects": True,
         "src_project_id": src_project.id,
     }
     g.Api.project.update_custom_data(dst_project.id, dst_custom_data)
@@ -71,22 +90,26 @@ def create_dst_dataset(
     return dst_dataset
 
 
-def download_yunet_model():
+def download_yunet_model(path: str = None):
     url = "https://github.com/opencv/opencv_zoo/raw/main/models/face_detection_yunet/face_detection_yunet_2023mar.onnx"
     file_name = "face_detection_yunet_2023mar.onnx"
-    model_path = Path(g.APP_DATA_DIR, "models", file_name)
-    model_path.parent.mkdir(parents=True, exist_ok=True)
-    r = requests.get(url, timeout=60)
-    with open(model_path, "wb") as f:
-        f.write(r.content)
-    return model_path.absolute().as_posix()
+    if path is None:
+        model_path = Path(g.APP_DATA_DIR, "models", file_name)
+    else:
+        model_path = Path(path)
+    if not model_path.exists():
+        model_path.parent.mkdir(parents=True, exist_ok=True)
+        r = requests.get(url, timeout=60)
+        with open(model_path, "wb") as f:
+            f.write(r.content)
+    return model_path.absolute()
 
 
 def get_yunet_model():
     if g.YUNET_MODEl is None:
         backend_id = cv2.dnn.DNN_BACKEND_OPENCV
         target_id = cv2.dnn.DNN_TARGET_CPU
-        model_path = download_yunet_model()
+        model_path = str(download_yunet_model())
 
         yunet = cv2.FaceDetectorYN.create(
             model=model_path,
@@ -121,15 +144,90 @@ def detect_faces_yunet(img: np.ndarray) -> np.ndarray:
     return res
 
 
-def _get_rectangles_mask(size, faces):
+def convert_bbox_to_coco(box: List) -> List:
+    x1, y1, x2, y2 = box
+    x = min(x1, x2)
+    y = min(y1, y2)
+    w = abs(x2 - x1)
+    h = abs(y2 - y1)
+    return [int(x), int(y), int(w), int(h)]
+
+
+def download_egoblur_model(path: str = None):
+    url = "https://scontent.fopo5-1.fna.fbcdn.net/m1/v/t6/An9sSpp_UpJ9wK4iapy8E1sWowJGvE3s-_npVcbow_FqLT4OJ0kiLsLOEnIUMC290kM3mfain4-Oomukg7ROXPYZr7YVpc8dJo-VYdOyneJ7HQNa8oi35HOE-H4yJ50wcKXc5eGeIg.zip/ego_blur_lp.zip?sdl=1&ccb=10-5&oh=00_AfAMHgC_-Bb7Bi3xA6rdCK5a8bTrzmQPTnL4vUt-gIN9zQ&oe=66073B3E&_nc_sid=5cb19f"
+    file_name_zip = "ego_blur_lp.zip"
+    file_name = "ego_blur_lp.jit"
+    if path is None:
+        model_path = Path(g.APP_DATA_DIR, "models", file_name)
+    else:
+        model_path = Path(path)
+    if model_path.is_file():
+        return model_path.absolute()
+
+    model_path_zip = model_path.with_name(file_name_zip)
+    model_path_zip.parent.mkdir(parents=True, exist_ok=True)
+    r = requests.get(url, timeout=60)
+    with open(model_path_zip, "wb") as f:
+        f.write(r.content)
+    with ZipFile(model_path_zip.absolute(), "r") as zip_ref:
+        zip_ref.extractall(model_path_zip.parent)
+
+    return model_path.absolute()
+
+
+def get_lp_egoblur():
+    if g.EGOBLUR_MODEl is None:
+        lp_model_path = str(download_egoblur_model())
+        lp_detector = torch.jit.load(lp_model_path, map_location="cpu").to(g.DEVICE)
+        lp_detector.eval()
+
+        g.EGOBLUR_MODEl = lp_detector
+    return g.EGOBLUR_MODEl
+
+
+def detect_lp_egoblur(
+    image: np.ndarray,
+):
+    model_score_threshold = g.STATE.threshold
+    nms_iou_threshold = 0.3
+    detector = get_lp_egoblur()
+    if len(image.shape) == 2:
+        image = cv2.cvtColor(image, cv2.COLOR_GRAY2BGR)
+    image_transposed = np.transpose(image, (2, 0, 1))
+    image_tensor = torch.from_numpy(image_transposed).to(device=g.DEVICE)
+
+    with torch.no_grad():
+        detections = detector(image_tensor)
+
+    boxes, _, scores, _ = detections  # returns boxes, labels, scores, dims
+
+    nms_keep_idx = torchvision.ops.nms(boxes, scores, nms_iou_threshold)
+    boxes = boxes[nms_keep_idx]
+    scores = scores[nms_keep_idx]
+
+    boxes = boxes.cpu().numpy()
+    scores = scores.cpu().numpy()
+
+    score_keep_idx = np.where(scores > model_score_threshold)[0]
+    boxes = boxes[score_keep_idx]
+
+    res = [
+        convert_bbox_to_coco(box.tolist()) + [float(score)]
+        for box, score in zip(boxes, scores)
+        if score > model_score_threshold
+    ]
+    return res
+
+
+def _get_rectangles_mask(size, objects):
     mask = np.zeros(size, dtype=np.uint8)
-    for x, y, w, h in faces:
+    for x, y, w, h in objects:
         mask[y : y + h, x : x + w] = 1
     return mask
 
 
-def blur_faces(img, faces, shape: str):
-    for x, y, w, h in faces:
+def blur_objects(img, objects, shape: str):
+    for x, y, w, h in objects:
         if shape == g.Shape.RECTANGLE:
             img[y : y + h, x : x + w] = cv2.blur(img[y : y + h, x : x + w], (h // 2, w // 2))
         elif shape == g.Shape.ELLIPSE:
@@ -152,20 +250,20 @@ def blur_faces(img, faces, shape: str):
     return img
 
 
-def obfuscate_faces(img: np.ndarray, faces: List, shape: str, method: str) -> np.ndarray:
+def obfuscate_objects(img: np.ndarray, objects: List, shape: str, method: str) -> np.ndarray:
     if shape not in g.AVAILABLE_SHAPES:
         raise ValueError(f"Invalid shape: {shape}")
     if method not in g.AVAILABLE_METHODS:
         raise ValueError(f"Invalid method: {method}")
 
     if method == g.Method.BLUR:
-        return blur_faces(img, faces, shape)
+        return blur_objects(img, objects, shape)
     else:
         if shape == g.Shape.RECTANGLE:
-            mask = _get_rectangles_mask(img.shape[:2], faces)
+            mask = _get_rectangles_mask(img.shape[:2], objects)
         elif shape == g.Shape.ELLIPSE:
             mask = np.zeros(img.shape[:2], dtype=np.uint8)
-            for x, y, w, h in faces:
+            for x, y, w, h in objects:
                 mask = cv2.ellipse(
                     mask,
                     (x + w // 2, y + h // 2),
@@ -179,46 +277,46 @@ def obfuscate_faces(img: np.ndarray, faces: List, shape: str, method: str) -> np
         return np.where(mask[:, :, None] == 1, 0, img)
 
 
-def update_annotation(dets, annotation: sly.Annotation, project_meta: sly.ProjectMeta):
-    face_obj_class = project_meta.get_obj_class(g.FACE_CLASS_NAME)
+def update_annotation(dets, class_name, annotation: sly.Annotation, project_meta: sly.ProjectMeta):
     conf_tag_meta = project_meta.get_tag_meta(g.CONFIDENCE_TAG_META_NAME)
 
+    obj_class = project_meta.get_obj_class(class_name)
     labels = []
     for det in dets:
         x, y, w, h, conf = det
 
         label = sly.Label(
             sly.Rectangle(y, x, y + h, x + w),
-            obj_class=face_obj_class,
+            obj_class=obj_class,
             tags=[sly.Tag(conf_tag_meta, conf)],
         )
         labels.append(label)
     return annotation.add_labels(labels)
 
 
-def face_coords_from_rectangle(rect: sly.Rectangle):
+def coords_from_rectangle(rect: sly.Rectangle):
     return (rect.left, rect.top, rect.right - rect.left, rect.bottom - rect.top)
 
 
-def faces_from_annotation(ann: sly.Annotation):
+def objects_from_annotation(ann: sly.Annotation):
     return [
-        face_coords_from_rectangle(label.geometry)
+        coords_from_rectangle(label.geometry)
         for label in ann.labels
         if (label.obj_class.name == g.FACE_CLASS_NAME and isinstance(label.geometry, sly.Rectangle))
     ]
 
 
-def filter_faces(faces: List):
-    faces = [tuple(face) for face in faces]
-    faces = set(faces)
-    return list(faces)
+def filter_objects(objects: List):
+    objects = [tuple(obj) for obj in objects]
+    objects = set(objects)
+    return list(objects)
 
 
 def run_images(
     src_dataset: sly.DatasetInfo,
     dst_dataset: sly.DatasetInfo,
     dst_project_meta: sly.ProjectMeta,
-    detector: Callable,
+    detectors: List[Callable],
     progress,
 ):
     for batch in g.Api.image.get_list_generator(src_dataset.id, batch_size=100):
@@ -233,20 +331,26 @@ def run_images(
         dst_nps = []
         for image_info in batch:
             img = g.Api.image.download_np(image_info.id)
-            dets = detector(img)
-            if g.STATE.should_save_detections:
-                anns_dict[image_info.id] = update_annotation(
-                    dets, anns_dict[image_info.id], dst_project_meta
+            for detector in detectors:
+                dets = detector(img)
+                class_name = (
+                    g.FACE_CLASS_NAME
+                    if detector.__name__ == "detect_faces_yunet"
+                    else g.LP_CLASS_NAME
                 )
-            if g.STATE.should_anonymize:
-                faces = [det[:4] for det in dets]
-                faces.extend(faces_from_annotation(anns_dict[image_info.id]))
-                img = obfuscate_faces(
-                    img,
-                    filter_faces(faces),
-                    g.STATE.obfuscate_shape,
-                    g.STATE.obfuscate_method,
-                )
+                if g.STATE.should_save_detections:
+                    anns_dict[image_info.id] = update_annotation(
+                        dets, class_name, anns_dict[image_info.id], dst_project_meta
+                    )
+                if g.STATE.should_anonymize:
+                    objects = [det[:4] for det in dets]
+                    objects.extend(objects_from_annotation(anns_dict[image_info.id]))
+                    img = obfuscate_objects(
+                        img,
+                        filter_objects(objects),
+                        g.STATE.obfuscate_shape,
+                        g.STATE.obfuscate_method,
+                    )
             dst_nps.append(img)
 
             progress.update(1)
@@ -255,7 +359,7 @@ def run_images(
         dst_img_metas = [
             {
                 **image_info.meta,
-                "anonymized_faces": g.STATE.should_anonymize,
+                "anonymized_objects": g.STATE.should_anonymize,
                 "src_image_id": image_info.id,
             }
             for image_info in batch
@@ -275,13 +379,13 @@ def run_videos(
     src_dataset: sly.DatasetInfo,
     dst_dataset: sly.DatasetInfo,
     dst_project_meta: sly.ProjectMeta,
-    detector: Callable,
+    detectors: List[Callable],
     progress,
 ):
     for batch in g.Api.video.get_list_generator(src_dataset.id, batch_size=1):
         for video in batch:
             dst_name = video.name
-            dst_video_meta = {**video.meta, "anonymized_faces": True, "src_video_id": video.id}
+            dst_video_meta = {**video.meta, "anonymized_objects": True, "src_video_id": video.id}
             video_path = os.path.join(g.APP_DATA_DIR, "videos", video.name)
             g.Api.video.download_path(
                 video.id,
@@ -303,8 +407,10 @@ def run_videos(
                 ret, frame = cap.read()
                 if not ret:
                     break
-                dets = detector(frame)
-                frame = obfuscate_faces(
+                dets = []
+                for detector in detectors:
+                    dets.extend(detector(frame))
+                frame = obfuscate_objects(
                     frame, [d[:4] for d in dets], g.STATE.obfuscate_shape, g.STATE.obfuscate_method
                 )
                 out.write(frame)
@@ -332,6 +438,20 @@ def get_total_items(datasets: List[sly.DatasetInfo], project_type) -> int:
         return s
 
 
+def download_models(dir: str = None):
+    download_yunet_model(Path(dir, "face_detection_yunet_2023mar.onnx"))
+    download_egoblur_model(Path(dir, "ego_blur_lp.jit"))
+
+
+def get_detectors():
+    if g.STATE.target == g.Model.YUNET:
+        return [detect_faces_yunet]
+    elif g.STATE.target == g.Model.EGOBLUR:
+        return [detect_lp_egoblur]
+    else:
+        return [detect_faces_yunet, detect_lp_egoblur]
+
+
 def main():
     src_project = g.Api.project.get_info_by_id(g.STATE.selected_project)
     dst_project = create_dst_project(src_project)
@@ -344,10 +464,8 @@ def main():
     run_func = run_images if src_project.type == str(sly.ProjectType.IMAGES) else run_videos
     total = get_total_items(datasets, src_project.type)
     with sly.tqdm.tqdm(total=total) as progress:
-        detector = detect_faces_yunet
-
+        detectors = get_detectors()
         for dataset in datasets:
             dst_dataset = create_dst_dataset(dataset, dst_project)
             dst_datasets.append(dst_dataset)
-
-            run_func(dataset, dst_dataset, dst_project_meta, detector, progress)
+            run_func(dataset, dst_dataset, dst_project_meta, detectors, progress)
